+++ J_Reactor_UI7.js
@@ -17,7 +17,7 @@ var Reactor = (function(api, $) {
 	/* unique identifier for this plugin... */
 	var uuid = '72acc6ea-f24d-11e8-bd87-74d4351650de';
 
-	var pluginVersion = '3.5hotfix-20070';
+	var pluginVersion = '3.5hotfix-20071';
 
 	var _UIVERSION = 20045;     /* must coincide with Lua core */